+++ web/store/index.ts
@@ -1,4 +1,8 @@
 export * from './user'
+export * from './authStore'
+export * from './subscriptionStore'
+export * from './uiStore'
+export * from './userConfigStore'
 export * from './chat'
 export * from './toasts'
 export * from './character'
